+++ ID_Payload.py
@@ -1,0 +1,18 @@
+class ScannedIDPayload(BaseModel):
+    uin:           str
+    name:          Optional[str] = None
+    dob:           Optional[str] = None
+    location1:     Optional[str] = None
+    location3:     Optional[str] = None
+    zone:          Optional[str] = None
+    postal_code:   Optional[str] = None
+    address_line1: Optional[str] = None
+    address_line2: Optional[str] = None
+    address_line3: Optional[str] = None
+
+    # ansure the UIN is not blank
+    @validator("uin")
+    def uin_must_not_be_empty(cls, v):
+        if not v.strip():
+            raise ValueError("UIN cannot be empty")
+        return v.strip()